On using roots and keyframes
Look at @keyframes wave { ... } to see how they move the penguin's arm! Cute.

Because of cascading, CSS variables are often defined in the :root element. You can think of the :root element as a container for your entire HTML document, in the same way that an html element is a container for the body element. By creating your variables in :root, they will be available throughout the whole web page.