Create blinking_stars
You can adjust the twinkling duration by changing animation-duration
You could also have done this by adjusting the @keyframes, but in this case the @keyframes uses @keyframes twinkle! And all of the stars have the SAME animation-name. Notice in this case that we have created classes instead of names for ids, and that's reflected in the divs at the bottom.